mix create_paddle_prod_plans (Plausible v0.0.1)

Utility for creating Paddle plans for production use.

Takes a single filename argument which should be of format input_plans_v*.json. That file should live in the /priv directory next to all other plans and it should contain the necessary information about the production plans to be created.

In order to create the input file:

Copy an existing plans_v*.json (latest recommended) into the new input_plans_v*.json file.

For every plan object:

Adjust the generation, limits, features, etc as desired

Replace monthly_product_id with a monthly_price (integer)

Replace yearly_product_id with a yearly_price (integer)

After this task is finished successfully, the plans will be created in Paddle with the prices given in the input file. With the creation, every plan gets an autoincremented ID in Paddle. We will then fetch those exact plans from Paddle in an API call and use their monthly and yearly product_id's to write plans_v*.json. It will be written taking the input file as the "template" and replacing the monthly/yearly prices with monthly/yearly product_id's.

The prices will be written into /priv/plan_prices.json (instead of the prod plans output file). Note that this separation is intentional - we only store prices locally to not rely on Paddle in the dev environment. Otherwise, Paddle is considered the "source of truth" of plan prices.

Usage example:

mix create_paddle_prod_plans input_plans_v5.json

Requirement 1: Replace the curl command

Unfortunately, there's no API in Paddle that would allow "bulk creating" plans - it has to be done through the UI. As a hack though, we can automate the process by copying the curl request with the help of browser devtools.

Therefore, this Mix.Task does not work out of the box and the actual curl command that it executes must be replaced by the developer. Here's how:

Access required to the production Paddle account

Navigate to https://vendors.paddle.com/subscriptions/plans. Chrome or Firefox recommended (need to copy a POST request as cURL in a later step)

Click the "+ New Plan" button (top right of the screen) to open the form

Open browser devtools, fill in the required fields and submit the form. No need to worry about the form fields since they're provided in this task (except _token) and they should work as long as nothing has changed.

Find the POST request from the "Network" tab and copy it as cURL

Come back here and paste it into the create_paddle_plan function

Replace the params within the string with the real params (these should be available in the function already)

Requirement 2: Paddle production credentials

You also need to obtain the Paddle credentials from prod environment and replace them into the module attributes. See @paddle_vendor_id and @paddle_vendor_auth_code. Those are needed to fetch the plans via an actual API call after the plans have been created in Paddle.

Functions

curl_quietly(cmd)

fetch_all_paddle_plans(url, paddle_credentials, page \\ 0, fetched \\ 0)

order_keys(plan)

run(list)

Callback implementation for Mix.Task.run/1.