mix create_paddle_sandbox_plans (Plausible v0.0.1)

Utility for creating Sandbox plans that are used on staging. The product of this task is a sandbox_plans_v*.json file matching with the production plans, just with the monthly/yearly product_id's of the sandbox plans.

In principle, this task works like Mix.Tasks.CreatePaddleProdPlans, with the following differences:

The filename argument should be the name of the JSON file containing the production plans (meaning that those should be created as the first step). No special "input file" required.

To copy the curl command from the browser, you need to add a plan from https://sandbox-vendors.paddle.com/subscriptions/plans. Everything else is the same - please see create_paddle_prod_plans.ex for instructions.

This task can be executed multiple times in a row - it will not create duplicates in Sandbox Paddle. On staging we can use a specific plan-naming structure to determine whether a plan has been created already. On prod we cannot do that since the plan names need to look nice.

No need to copy paddle API credentials.

Usage example:

mix create_paddle_sandbox_plans plans_v5.json

Functions

run(list)

Callback implementation for Mix.Task.run/1.