Plausible.Annotations.Annotation (Plausible v0.0.1)

Schema for annotations. Annotations are notes attached to a point on the graph.

Annotations can have two granularities, date or minute.

To create date granularity annotations, only the date part must be specified for datetime field. It's interpreted to be for that date, no matter what happens with the site timezone.

Examples:

{"granularity": "date", "datetime": "2026-06-30", ...}.

To create minute granularity annotations, the full datetime needs to be specified for datetime field. It can be set in local time, as a naive datetime string ("2026-06-29 10:00:00"), in which case it's interpreted as that datetime in the site's timezone and stored as that particular moment in UTC. Alternatively, it can be sent with TZ information ("2026-05-31T12:00:00Z", "2026-05-31T10:00:00-02:00"), in which case it's interpreted to be that particular moment and stored as that in UTC.

Examples:

{"granularity": "minute", "datetime": "2026-06-29 10:00:00", ...}

{"granularity": "minute", "datetime": "2026-06-29T12:00:00Z", ...}

Types

t()

@type t() :: %Plausible.Annotations.Annotation{
  __meta__: term(),
  date: term(),
  datetime: term(),
  granularity: term(),
  id: term(),
  inserted_at: term(),
  note: term(),
  owner: term(),
  owner_id: term(),
  site: term(),
  site_id: term(),
  type: term(),
  updated_at: term()
}

Functions

changeset(annotation, attrs, site_timezone)

create_changeset(attrs, site, owner)

serialize_date_granularity_datetime(date)

update_changeset(annotation, attrs, owner)