Plausible.Auth.SSO.Domain (Plausible v0.0.1)

Once SSO integration is initiated, it's possible to start allow-listing domains for it, in parallel with finalizing the setup on IdP's end.

Each pending domain should be periodically checked for ownership verification by testing for presence of TXT record, meta tag or URL. The moment whichever of them succeeds first, the domain is marked as verified with method and timestamp recorded.

Types

t()

@type t() :: %Plausible.Auth.SSO.Domain{
  __meta__: term(),
  domain: term(),
  id: term(),
  identifier: term(),
  inserted_at: term(),
  last_verified_at: term(),
  sso_integration: term(),
  sso_integration_id: term(),
  status: term(),
  updated_at: term(),
  verified_via: term()
}

verification_method()

@type verification_method() :: :meta_tag | :url | :dns_txt

Functions

create_changeset(integration, domain)

@spec create_changeset(Plausible.Auth.SSO.Integration.t(), String.t() | nil) ::
  Ecto.Changeset.t()

unverified_changeset(sso_domain, status \\ Status.in_progress(), now \\ NaiveDateTime.utc_now(:second))

@spec unverified_changeset(t(), atom(), NaiveDateTime.t()) :: Ecto.Changeset.t()

valid_domain?(domain)

@spec valid_domain?(String.t()) :: boolean()

verification_methods()

@spec verification_methods() :: [verification_method()]

verified_changeset(sso_domain, method, now \\ NaiveDateTime.utc_now(:second))

@spec verified_changeset(t(), verification_method(), NaiveDateTime.t()) ::
  Ecto.Changeset.t()