Plausible.Billing.Plans (Plausible v0.0.1)

find(product_id_or_subscription, scope \\ all())

@spec find(String.t() | Plausible.Billing.Subscription.t(), [
  Plausible.Billing.Plan.t()
]) ::
  Plausible.Billing.Plan.t() | nil

@spec find(nil, any()) :: nil

Finds a plan by product ID.

Returns nil when plan can't be found.

for_user(user)

@spec for_user(Plausible.Auth.User.t()) :: [Plausible.Billing.Plan.t()]

Returns a list of plans available for the user to choose.

As new versions of plans are introduced, users who were on old plans can still choose from old plans.

get_subscription_plan(subscription)

subscription_interval(subscription)

suggest(user, usage_during_cycle)

@spec suggest(Plausible.Auth.User.t(), non_neg_integer()) ::
  Plausible.Billing.Plan.t()

Returns the most appropriate plan for a user based on their usage during a given cycle.

If the usage during the cycle exceeds the enterprise-level threshold, or if the user already belongs to an enterprise plan, it suggests the :enterprise plan.

Otherwise, it recommends the plan where the cycle usage falls just under the plan's limit from the available options for the user.

yearly_product_ids()

@spec yearly_product_ids() :: [String.t()]

List yearly plans product IDs.