Plausible.Event.SystemEvents (Plausible v0.0.1)

System events are events that require at least one form of special treatment by the application. They are distinguished by event name.

For some system events, the tracking API may reject events that don't match the expected format (e.g. engagement events without scroll depth or engagement time defined).

For other system events, it may accept the event (e.g. "Outbound Link: Click" with the url prop not set).

System events may have corresponding system-managed goals, in which case the goal name and event name will be the same (e.g. "Outbound Link: Click" goal is managed by the application). The system will only be able to manage these goals properly if they are not renamed by the user.

Functions

events()

events_with_engagement_props()

events_with_interactive_always_true()

events_with_path_prop()

events_with_url_prop()

special_events_for_prop_key(binary)

@spec special_events_for_prop_key(String.t()) :: [String.t()]

sync_props_path_with_pathname?(event_name, props_in_request)

@spec sync_props_path_with_pathname?(String.t(), [{String.t(), String.t()}]) ::
  boolean()

Checks if the event name is for a system event / system goal that should have the event.props.path synced with the event.pathname property.

Examples

iex> sync_props_path_with_pathname?("404", [{"path", "/foo"}]) false

Note: Should not override event.props.path if it is set deliberately to nil iex> sync_props_path_with_pathname?("404", [{"path", nil}]) false

iex> sync_props_path_with_pathname?("404", [{"other", "value"}]) true

iex> sync_props_path_with_pathname?("404", []) true