Plausible.Google.Api (Plausible v0.0.1)

google_analytics_view()

@type google_analytics_view() :: {view_name :: String.t(), view_id :: String.t()}

import_auth()

@type import_auth() ::
  {access_token :: String.t(), refresh_token :: String.t(),
   expires_at :: String.t()}

fetch_and_persist(site, report_request, opts \\ [])

@spec fetch_and_persist(
  Plausible.Site.t(),
  Plausible.Google.ReportRequest.t(),
  Keyword.t()
) ::
  :ok | {:error, term()}

fetch_stats(site, map, limit)

fetch_verified_properties(auth)

get_view(access_token, lookup_id)

@spec get_view(access_token :: String.t(), lookup_id :: String.t()) ::
  {:ok, google_analytics_view()} | {:ok, nil} | {:error, term()}

Returns a single Google Analytics view if the user has access to it.

import_analytics(site, date_range, view_id, auth)

@spec import_analytics(Plausible.Site.t(), Date.Range.t(), String.t(), import_auth()) ::
  :ok | {:error, term()}

Imports stats from a Google Analytics UA view to a Plausible site.

This function fetches Google Analytics reports in batches of 7500 per request. The batches are then buffered to Clickhouse by the Plausible.Google.Buffer process.

Requests to Google Analytics can fail, and are retried at most 5 times with an exponential backoff. Returns :ok when importing has finished or {:error, term()} when a request to GA failed too many times.

Useful links:

Feature documentation

GA API reference

GA Dimensions reference

import_authorize_url(site_id, redirect_to)

list_views(access_token)

@spec list_views(access_token :: String.t()) ::
  {:ok, %{required(hostname :: String.t()) => [google_analytics_view()]}}
  | {:error, term()}

Lists Google Analytics views grouped by hostname.

search_console_authorize_url(site_id, redirect_to)