Plausible.Google.Buffer (Plausible v0.0.1)

This GenServer inserts records into Clickhouse imported_* tables. Multiple buffers are automatically created for each table. Records are flushed when the table buffer reaches the maximum size, defined by max_buffer_size/0.

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

flush(pid, timeout \\ :infinity)

Flushes all table buffers to Clickhouse.

handle_continue(arg, state)

Callback implementation for GenServer.handle_continue/2.

init(opts)

Callback implementation for GenServer.init/1.

insert_many(pid, table_name, records)

@spec insert_many(pid(), term(), [map()]) :: :ok

Puts the given records into the table buffer.

size(pid, table_name)

@spec size(pid(), term()) :: non_neg_integer()

Returns the total count of items in the given table buffer.

start_link()

stop(pid)