Plausible.Ingestion.Acquisition (Plausible v0.0.1)

This module is responsible for figuring out acquisition channel from event referrer_source.

Acquisition channel is the marketing channel where people come from and convert and help users to understand and improve their marketing flow.

Note it uses priv/ga4-source-categories.csv as a source, which comes from https://support.google.com/analytics/answer/9756891?hl=en.

Notable differences from GA4 that have been implemented just for Plausible:

The @custom_source_categories module attribute contains a list of custom source categories that we have manually added based on our own judgement and user feedback. For example we treat AI tools (ChatGPT, Perplexity) as search engines.

Google is in a privileged position to analyze paid traffic from within their own network. The biggest use-case is auto-tagged adwords campaigns. We do our best by categorizing as paid search when source is Google and the url has gclid parameter. Same for source Bing and msclkid url parameter.

The @paid_sources module attribute in Plausible.Ingestion.Source contains a list of utm_sources that we will automatically categorize as paid traffic regardless of the medium. Examples are yt-ads, facebook_ad, adwords, etc. See also: Plausible.Ingestion.Source.paid_source?/1

Functions

get_channel(source, utm_medium, utm_campaign, utm_source, click_id_param)

source_categories()