Plausible.Site.Cache (Plausible v0.0.1)

A sites by domain caching interface.

Serves as a thin wrapper around Cachex, but the underlying implementation can be transparently swapped.

Even though the Cachex process is started, cache access is disabled during tests via the :sites_by_domain_cache_enabled application env key. This can be overriden on case by case basis, using the child specs options.

When Cache is disabled via application env, the get/1 function falls back to pure database lookups. This should help with introducing cached lookups in existing code, so that no existing tests should break.

To differentiate cached Site structs from those retrieved directly from the database, a virtual schema field from_cache?: true is set. This indicates the Plausible.Site struct is incomplete in comparison to its database counterpart -- to spare bandwidth and query execution time, only selected database columns are retrieved and cached.

The @cached_schema_fields attribute defines the list of DB columns queried on cache pre-fill.

Also see tests for more comprehensive examples.

Functions

child_spec(opts)

@spec child_spec(Keyword.t()) :: Supervisor.child_spec()

enabled?()

get(domain, opts \\ [])

@spec get(String.t(), Keyword.t()) :: nil | Plausible.Site.t()

hit_rate(cache_name \\ :sites_by_domain)

name()

prefill(opts)

@spec prefill(Keyword.t()) :: :ok

size(cache_name \\ :sites_by_domain)