Plausible.Stats.Fragments (Plausible v0.0.1)

bounce_rate()
(macro)

coalesce_string(fieldA, fieldB)
(macro)

sample_percent()
(macro)

to_timezone(date, timezone)
(macro)

Converts time or date and time to the specified timezone.

Reference: https://clickhouse.com/docs/en/sql-reference/functions/date-time-functions/#totimezone

total()
(macro)

uniq(user_id)
(macro)

visit_duration()
(macro)

weekstart_not_before(date, not_before)
(macro)

Returns the weekstart for date. If the weekstart is before the not_before boundary, not_before is returned.

Examples

In this pseudo-code example, the fragment returns the weekstart. The not_before boundary is set to the past Saturday, which is before the weekstart, therefore the cap does not apply.

iex> this_wednesday = ~D[2022-11-09] ...> past_saturday = ~D[2022-11-05] ...> weekstart_not_before(this_wednesday, past_saturday) ~D[2022-11-07]

In this other example, the fragment returns Tuesday and not the weekstart. The not_before boundary is set to Tuesday, which is past the weekstart, therefore the cap applies.

iex> this_wednesday = ~D[2022-11-09] ...> this_tuesday = ~D[2022-11-08] ...> weekstart_not_before(this_wednesday, this_tuesday) ~D[2022-11-08]

weekstart_not_before(date, not_before, timezone)
(macro)

Same as Plausible.Stats.Fragments.weekstart_not_before/2 but converts dates to the specified timezone.