Plausible.Stats.Goal.Revenue (Plausible v0.0.1)

Revenue specific functions for the stats scope

Functions

available?(site)

format_revenue_metric(value, query, dimension_values)

preload(site, preloaded_goals, metrics, dimensions)

Preloads revenue currencies for a query. Used when parsing the query.

Returns tuple containing revenue warning (if set, no revenue metrics should be calculated) and revenue currencies map.

Assumptions and business logic:

Goals are already filtered according to query filters and dimensions

If there's a single currency involved, return map containing the default

If there's a breakdown by event:goal we return all the relevant currencies as a map

If filtering by multiple different currencies without event:goal breakdown empty map is returned

If user has no access or preloading is not needed, empty map is returned

The resulting data structure is attached to a Query and used below in format_revenue_metric/3.

requested?(metrics)

revenue_metrics()