Plausible.Stats.Goals (Plausible v0.0.1)

Stats code related to filtering and grouping by goals.

Types

goal_join_data()

@type goal_join_data() :: %{
  indices: [non_neg_integer()],
  types: [String.t()],
  event_names_imports: [String.t()],
  event_names_by_type: [String.t()],
  page_regexes: [String.t()],
  scroll_thresholds: [non_neg_integer()],
  custom_props_keys: [[String.t()]],
  custom_props_values: [[String.t()]]
}

Functions

add_filter(query, filter, opts \\ [])

Translates an event:goal filter into SQL. Similarly to other add_filter clauses in Plausible.Stats.SQL.WhereBuilder, returns an Ecto.Query.dynamic expression.

Compared to other dimensions, filtering by goals works differently. First, we expect all goals to be preloaded into the query - this list of goals is what actually gets filtered by operator and clauses.

With the resulting filtered list of goals, we build conditions for whether that goal was completed or not, and join those conditions with a logical OR.

Options

imported? - when true, builds conditions on the page db field rather than pathname, and also skips the e.name == "pageview" check.

goal_condition(goal, imported? \\ false)

goal_join_data(query)

@spec goal_join_data(Plausible.Stats.Query.t()) :: goal_join_data()

Returns data needed to perform a GROUP BY on goals in an ecto query.

page_path_db_field(imported?)

preload_needed_goals(site, dimensions, filters)

Preloads goals data if needed for query-building and related work.

toplevel_scroll_goal_filters?(query)