Plausible.Stats.Imported (Plausible v0.0.1)

Functions

merge_imported(q, site, query)

merge_imported_city_suggestions(native_q, site, query)

merge_imported_country_suggestions(native_q, site, query)

merge_imported_filter_suggestions(native_q, site, query, filter_name, filter_query)

merge_imported_region_suggestions(native_q, site, query)

schema_supports_query?(query)

Returns a boolean indicating whether the combination of filters and breakdown property is possible to query from the imported tables.

Usually, when no filters are used, the imported schema supports the query. There is one exception though - breakdown by a custom property. We are currently importing only two custom properties - url and path. Both these properties can only be used with their special goal filter (see Plausible.Goals.SystemGoals).

total_imported_visitors(site, query)