Plausible.Stats.Interval (Plausible v0.0.1)

Collection of functions to work with intervals.

The interval of a query defines the granularity of the data. You can think of it as a GROUP BY clause. Possible values are minute, hour, day, week, and month.

Types

opt()

@type opt() :: {:site, Plausible.Site.t()} | {:from, Date.t()}

opts()

@type opts() :: [opt()]

t()

@type t() :: String.t()

Functions

default_for_query(query)

@spec default_for_query(Plausible.Stats.Query.t()) :: t()

Returns the suggested interval (i.e. the time dimension) for the given query.

list()

@spec list() :: [t()]

valid?(interval)

@spec valid?(term()) :: boolean()

valid_by_period(opts \\ [])

@spec valid_by_period(opts()) :: map()

valid_for_period?(period, interval, opts \\ [])

@spec valid_for_period?(period(), t(), opts()) :: boolean()

Returns whether the given interval is valid for a time period.

Intervals longer than periods are not supported, e.g. current month stats with a month interval, or today stats with a week interval.

There are two dynamic states:

custom period is only applicable with month or week intervals, if the opts[:from] and opts[:to] range difference exceeds 12 months

all period's interval options depend on particular site's stats_start_date

daily interval is excluded if the all-time range exceeds 12 months