Plausible.Stats.Legacy.TimeOnPage (Plausible v0.0.1)

Calculation methods for legacy time_on_page metric. Legacy calculation methods are used when site does not have engagement data for the requested dates.

Query include.legacy_time_on_page_cutoff is used to determine what time range to use for the legacy time_on_page calculations.

Legacy metric is not exposed in the public API.

Functions

can_merge_legacy_time_on_page?(query)

merge_legacy_time_on_page(q, query)