Plausible.Stats.SQL.Expression (Plausible v0.0.1)

This module is responsible for generating SQL/Ecto expressions for dimensions and metrics used in query SELECT statement.

Each dimension and metric is tagged with with selected_as for easier usage down the line.

Functions

event_goal_join(goal_join_data)
(macro)

The fragment matches events to goals by:

Checking if the pathname matches the goal's page regex pattern

Verifying the event name matches the expected name for the goal type

Validating scroll depth is within threshold (for scroll goals)

Ensuring all custom properties match (if any are defined on the goal)

Returns an array of goal indices that the event matches.

event_metric(unknown, query)

select_dimension(q, key, arg3, arg4, query)

select_dimension_from_join(q, key, arg3)

select_dimension_internal(q, arg2)

session_metric(atom, query)