Plausible.Teams (Plausible v0.0.1)

Core context of teams.

Functions

get_by_owner(user)

get_or_create(user)

Get or create user's team.

If the user has no non-guest membership yet, an implicit "My Team" team is created with them as an owner.

If the user already has an owner membership in an existing team, that team is returned.

If the user has a non-guest membership other than owner, :no_team error is returned.

last_subscription_join_query()

last_subscription_query()

load_for_site(site)

Create (when necessary) and load team relation for provided site.

Used for sync logic to work smoothly during transitional period.

on_trial?(team)

@spec on_trial?(Plausible.Teams.Team.t() | nil) :: boolean()

owned_sites(team)

owned_sites_ids(team)

read_team_schemas?(user)

sync_team(user)

trial_days_left(team)

@spec trial_days_left(Plausible.Teams.Team.t()) :: integer()

with_subscription(team)