Plausible.Teams.GracePeriod (Plausible v0.0.1)

This embedded schema stores information about the account locking grace period.

Teams are given this 7-day grace period to upgrade their account after outgrowing their subscriptions. The actual account locking happens in background with Plausible.Workers.LockSites.

The grace period can also be manual, without an end date, being controlled manually from the CRM, and not by the background site locker job. This is useful for enterprise subscriptions.

Types

t()

@type t() :: %Plausible.Teams.GracePeriod{
  end_date: Date.t() | nil,
  id: term(),
  is_over: boolean(),
  manual_lock: boolean()
}

Functions

active?(team)

@spec active?(Plausible.Teams.Team.t() | nil) :: boolean()

Returns whether the grace period is still active for a Team. Defaults to false if the team is nil or there is no grace period.

end_changeset(team)

@spec end_changeset(Plausible.Teams.Team.t()) :: Ecto.Changeset.t()

Ends an existing grace period by setting teams.grace_period.is_over to true. This means the grace period has expired.

expired?(team)

@spec expired?(Plausible.Teams.Team.t() | nil) :: boolean()

Returns whether the grace period has already expired for a Team. Defaults to false if the team is nil or there is no grace period.

remove_changeset(team)

@spec remove_changeset(Plausible.Teams.Team.t()) :: Ecto.Changeset.t()

Removes the grace period from the Team completely.

start_changeset(team)

@spec start_changeset(Plausible.Teams.Team.t()) :: Ecto.Changeset.t()

Starts a account locking grace period of 7 days by changing the Team struct.

start_manual_lock_changeset(team)

@spec start_manual_lock_changeset(Plausible.Teams.Team.t()) :: Ecto.Changeset.t()

Starts a manual account locking grace period by changing the Team struct. Manual locking means the grace period can only be removed manually from the CRM.