Plausible.Teams.Management.Layout (Plausible v0.0.1)

Abstraction for team membership(s) layout - provides a high level CRUD for setting up team memberships, including invitations. Persisting the layout, effectively takes care of delegating the operations to specialized services and sending out e-mail notifications on success, if need be. To be used in UIs allowing team memberships adjustments.

Types

t()

@type t() :: %{required(String.t()) => Plausible.Teams.Management.Layout.Entry.t()}

Functions

active_count(layout)

@spec active_count(t()) :: non_neg_integer()

build_by_email(entities)

@spec build_by_email([
  Plausible.Teams.Invitation.t() | Plausible.Teams.Membership.t()
]) :: t()

has_guests?(layout)

@spec has_guests?(t()) :: boolean()

init(team)

@spec init(Plausible.Teams.Team.t()) :: t()

owners_count(layout)

@spec owners_count(t()) :: non_neg_integer()

persist(layout, context)

@spec persist(t(), %{
  current_user: Plausible.Auth.User.t(),
  current_team: Plausible.Teams.Team.t()
}) ::
  {:ok, integer()} | {:error, any()}

persist_entry(entry, context, acc)

removable?(layout, email)

@spec removable?(t(), String.t()) :: boolean()

schedule_delete(layout, email)

@spec schedule_delete(t(), String.t()) :: t()

schedule_send(layout, email, role, entry_attrs \\ [])

@spec schedule_send(t(), String.t(), atom(), Keyword.t()) :: t()

sorted_for_display(layout)

@spec sorted_for_display(t()) :: [
  {String.t(), Plausible.Teams.Management.Layout.Entry.t()}
]

update_role(layout, email, role)

@spec update_role(t(), String.t(), atom()) :: t()

verify_removable(layout, email)

@spec verify_removable(t(), String.t()) :: :ok | {:error, String.t()}