Plausible.Test.Support.HTML (Plausible v0.0.1)

LazyHTML wrappers to help make assertions about HTML/DOM structures

Functions

attr_defined?(html, element, attr)

class_of_element(html, element)

elem_count(html, selector)

element_exists?(html, selector)

find(html, selector)

find_portal_template(html, id)

When using liveview <.portal> element, it renders a <template> that gets rendered into DOM by JS. In order to make assertions about portal contents, we need to find the <template> element and get its contents. Unfortunately <template> elements are not treated like standard elements by LazyHTML. Functions like LazyHTML.text() or LazyHTML.query() return nothing for <template> elements. So this function tricks LazyHTML by:

1. Finding the template element
2. Transforming it into a <div> so it can be used like a normal LazyHTML node

)

form_exists?(html, action_path)

name_of(element)

submit_button(html, form)

text(element)

text_of_attr(element, attr)

text_of_attr(html, selector, attr)

text_of_element(html, selector)