Plausible.Verification.Check behaviour (Plausible v0.0.1)

Behaviour to be implemented by specific site verification checks. report_progress_as() doesn't necessarily reflect the actual check description, it serves as a user-facing message grouping mechanism, to prevent frequent message flashing when checks rotate often. Each check operates on state() and is expected to return it, optionally modified, by all means. perform_safe/1 is used to guarantee no exceptions are thrown by faulty implementations, not to interrupt LiveView.

Types

state()

@type state() :: Plausible.Verification.State.t()

Callbacks

perform(state)

@callback perform(state()) :: state()

report_progress_as()

@callback report_progress_as() :: String.t()