Plausible.Workers.ScheduleEmailReports (Plausible v0.0.1)

first_of_month_9am(timezone)

monday_9am(timezone)

perform(job)

Email reports should be sent on Monday at 9am according to the timezone of a site. This job runs every day at midnight to ensure that all sites have a scheduled job for email reports.