PlausibleWeb.AuthView (Plausible v0.0.1)

__resource__()

The resource name, as an atom, for this view

activate.html(assigns)

base_domain()

delimit_integer(number)

format_invoices(invoice_list)

invitation_expired.html(assigns)

login_form.html(assigns)

new_api_key.html(assigns)

password_form.html(assigns)

password_reset_form.html(assigns)

password_reset_request_form.html(assigns)

password_reset_request_success.html(assigns)

plausible_url()

present_subscription_status(status)

register_form.html(assigns)

register_from_invitation_form.html(assigns)

register_success.html(assigns)

render(template, assigns \\ %{})

Renders the given template locally.

subscription_colors(_)

subscription_interval(subscription)

subscription_quota(subscription)

template_not_found(template, assigns)

@spec template_not_found(binary(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

user_settings.html(assigns)