PlausibleWeb.Components.Generic (Plausible v0.0.1)

Generic reusable components

Functions

accordion_item(assigns)

Attributes

id (:string) (required)

title (:string) (required)

open_by_default (:boolean) - Defaults to false.

title_class (:string) - Defaults to "".

Slots

inner_block (required)

accordion_menu(assigns)

Slots

inner_block (required)

button(assigns)

Attributes

type (:string) - Defaults to "button".

theme (:string) - Defaults to "primary".

class (:string) - Defaults to "".

disabled (:boolean) - Defaults to false.

mt? (:boolean) - Defaults to true.

Global attributes are accepted. Supports all globals plus: ["name"].

Slots

inner_block

button_link(assigns)

Attributes

href (:string) (required)

class (:string) - Defaults to "".

theme (:string) - Defaults to "primary".

disabled (:boolean) - Defaults to false.

method (:string) - Defaults to "get".

mt? (:boolean) - Defaults to true.

Global attributes are accepted.

Slots

inner_block

delete_button(assigns)

Attributes

href (:string) - Defaults to nil.

class (:string) - Defaults to "".

icon (:atom) - Defaults to :trash.

Global attributes are accepted. Supports all globals plus: ["method", "disabled"].

disclosure(assigns)

Slots

inner_block (required)

disclosure_button(assigns)

Attributes

class (:any) - Defaults to nil.

Slots

inner_block (required)

disclosure_panel(assigns)

Slots

inner_block (required)

docs_info(assigns)

Attributes

slug (:string) (required)

class (:string) - Defaults to nil.

dropdown(assigns)

Attributes

class (:string) - Defaults to "".

id (:string) - Defaults to nil.

Slots

button (required) - Accepts attributes:

class (:string)

menu (required) - Accepts attributes:

class (:string)

dropdown_divider(assigns)

dropdown_item(assigns)

Attributes

href (:string)

class (:string) - Defaults to "".

id (:string) - Defaults to nil.

new_tab (:boolean) - Defaults to false.

disabled (:boolean) - Defaults to false.

Global attributes are accepted. Supports all globals plus: ["method"].

Slots

inner_block (required)

dynamic_icon(assigns)

Attributes

name (:atom) (required)

outline (:boolean) - Defaults to true.

solid (:boolean) - Defaults to false.

mini (:boolean) - Defaults to false.

Global attributes are accepted. Supports all globals plus: ["fill", "stroke", "stroke-width"].

edit_button(assigns)

Attributes

href (:string) - Defaults to nil.

class (:string) - Defaults to "".

Global attributes are accepted. Supports all globals plus: ["method", "disabled"].

filter_bar(assigns)

Attributes

filter_text (:string) - Defaults to "".

placeholder (:string) - Defaults to "".

filtering_enabled? (:boolean) - Defaults to true.

Slots

inner_block

focus_box(assigns)

Attributes

padding? (:boolean) - Defaults to true.

Global attributes are accepted.

Slots

title

subtitle

inner_block (required)

footer

focus_list(assigns)

Slots

item (required)

h2(assigns)

Attributes

class (:any) - Defaults to nil.

Global attributes are accepted.

Slots

inner_block (required)

highlighted(assigns)

Slots

inner_block (required)

icon_button(assigns)

Attributes

href (:string)

icon_name (:atom)

theme (:string) - Defaults to "default".

class (:string) - Defaults to "".

icon_class (:string) - Defaults to "".

size (:string) - Defaults to "4".

Global attributes are accepted. Supports all globals plus: ["method", "disabled"].

Slots

inner_block

notice(assigns)

Attributes

title (:any) - Defaults to nil.

theme (:atom) - Defaults to :yellow.

dismissable_id (:any) - Defaults to nil.

class (:string) - Defaults to "".

Global attributes are accepted.

Slots

inner_block

pill(assigns)

Attributes

class (:string) - Defaults to "".

color (:atom) - Defaults to :gray. Must be one of :gray, :indigo, :yellow, or :green.

Global attributes are accepted.

Slots

inner_block (required)

settings_tiles(assigns)

spinner(assigns)

Attributes

class (:any) - Defaults to "".

Global attributes are accepted.

styled_link(assigns)

Attributes

href (:string) - Defaults to "#".

new_tab (:boolean) - Defaults to false.

class (:string) - Defaults to "".

method (:string) - Defaults to "get".

Global attributes are accepted. Supports all globals plus: ["patch"].

Slots

inner_block

table(assigns)

Attributes

width (:string) - Defaults to "min-w-full".

rows (:list) - Defaults to [].

row_attrs (:any) - Defaults to nil.

Global attributes are accepted.

Slots

thead

tbody (required)

inner_block

td(assigns)

Attributes

truncate (:boolean) - Defaults to false.

max_width (:string) - Defaults to "".

height (:string) - Defaults to "".

class (:string) - Defaults to "".

actions (:boolean) - Defaults to nil.

hide_on_mobile (:boolean) - Defaults to nil.

Global attributes are accepted.

Slots

inner_block (required)

th(assigns)

Attributes

invisible (:boolean) - Defaults to false.

hide_on_mobile (:boolean) - Defaults to nil.

Slots

inner_block (required)

tile(assigns)

Attributes

docs (:string) - Defaults to nil.

feature_mod (:atom) - Defaults to nil.

feature_toggle? (:boolean) - Defaults to false.

current_team (:any) - Defaults to nil.

current_user (:any) - Defaults to nil.

site (:any) - Defaults to nil.

conn (:any) - Defaults to nil.

show_content? (:boolean) - Defaults to true.

Slots

inner_block (required)

title (required)

subtitle

title(assigns)

Attributes

class (:any) - Defaults to nil.

Slots

inner_block (required)

toggle_field(assigns)

Renders toggle input with a label. Clicking the label also toggles the toggle. Needs :js_active_var that controls toggle state. Set this outside this component with x-data="{ <variable name>: <initial state> }" Can be configured to always show a description of the field / help text :help_text, or only show the help text when the toggle is activated :show_help_text_only_when_active?.

Attributes

id (:string) (required)

js_active_var (:string) (required)

id_suffix (:string) - Defaults to "".

disabled (:boolean) - Defaults to false.

label (:string) (required)

help_text (:string) - Defaults to nil.

show_help_text_only_when_active? (:boolean) - Defaults to false.

mt? (:boolean) - Defaults to true.

Global attributes are accepted.

toggle_submit(assigns)

Attributes

set_to (:boolean) - Defaults to false.

disabled? (:boolean) - Defaults to false.

Slots

inner_block (required)

toggle_switch(assigns)

Renders toggle input. Needs :js_active_var that controls toggle state. Set this outside this component with x-data="{ <variable name>: <initial state> }".

### Examples

  <div x-data="{ showGoals: false }>
    <.toggle_switch id="show_goals" js_active_var="showGoals" />
  </div>

## Attributes

id (:string) (required) js_active_var (:string) (required) id_suffix (:string) - Defaults to "". disabled (:boolean) - Defaults to false. * Global attributes are accepted.

tooltip(assigns)

Attributes

sticky? (:boolean) - Defaults to true.

enabled? (:boolean) - Defaults to true.

centered? (:boolean) - Defaults to false.

testid (:string) - Defaults to nil.

Slots

inner_block (required)

tooltip_content (required)

unstyled_link(assigns)

Attributes

href (:string) (required)

new_tab (:boolean) - Defaults to false.

class (:string) - Defaults to "".

method (:string) - Defaults to "get".

Global attributes are accepted.

Slots

inner_block

upgrade(assigns)

Attributes

title (:any) - Defaults to "".

class (:string) - Defaults to "".

Global attributes are accepted.

Slots

icon (required)

inner_block