PlausibleWeb.Components.Site.Feature (Plausible v0.0.1)

Phoenix Component for rendering a user-facing feature toggle capable of flipping booleans in Plausible.Site via the toggle_feature controller action.

__resource__()

The resource name, as an atom, for this view

render(template, assigns \\ %{})

Renders the given template locally.

target(site, setting, conn, set_to)

template_not_found(template, assigns)

@spec template_not_found(binary(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

toggle(assigns)

Attributes

site (Plausible.Site) (required)

feature_mod (:atom) (required) - Must be one of Plausible.Billing.Feature.Goals, Plausible.Billing.Feature.StatsAPI, Plausible.Billing.Feature.Props, Plausible.Billing.Feature.Funnels, or Plausible.Billing.Feature.RevenueGoals.

conn (Plug.Conn) (required)

Slots

inner_block