PlausibleWeb.Flows (Plausible v0.0.1)

Named identifiers and documentation for the onboarding/installation flows.

Functions

add_site_step()

domain_change()

The flow of an existing site domain getting changed (triggered from site settings):

Domain name change form

Domain change success info page (possible further actions required)

Back to site settings

installation_step()

invitation()

The flow of accepting a team/site invitation as a new user:

Register form

Activate account (email confirmation)

Land on the /sites page

onboarding_steps()

The steps shown in the onboarding flow's progress indicator.

provisioning()

The flow of an already logged in user adding another site to their team:

Add site info

Installation screen

End up either on the dashboard (with verification kicked off automatically) or the /sites page, depending on whether installation was skipped or not

register()

The flow of creating a new account from scratch:

Register form

Activate account (email confirmation)

...optional onboarding steps (see onboarding_steps/0)

End up either on the dashboard (with verification kicked off automatically) or the /sites page, depending on whether a site was setup or not.

review()

The flow of an existing user reviewing/re-checking their already-installed tracking script (triggered from site settings):

Installation screen, pre-filled with the earlier installation method

End up on the dashboard (with verification kicked off automatically)