PlausibleWeb.LayoutView (Plausible v0.0.1)

Functions

__resource__()

The resource name, as an atom, for this view

app.html(assigns)

base_email.html(assigns)

base_error.html(assigns)

embedded.html(assigns)

feedback_link(user)

flat_settings_options(conn)

grace_period_end(user)

home_dest(conn)

is_current_tab(conn, tab)

logo_path(filename)

plausible_url()

priority_email.html(assigns)

render(template, assigns \\ %{})

Renders the given template locally.

render_layout(layout, assigns, list)

http://blog.plataformatec.com.br/2018/05/nested-layouts-with-phoenix/

settings.html(assigns)

settings_tabs(conn)

site_settings.html(assigns)

template_not_found(template, assigns)

@spec template_not_found(binary(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

trial_notificaton(user)

websocket_url()