PlausibleWeb.Live.Components.Form (Plausible v0.0.1)

Generic components stolen from mix phx.new templates

error(assigns)

Generates a generic error message.

Slots

inner_block (required)

input(assigns)

Renders an input with label and error messages.

A Phoenix.HTML.FormField may be passed as argument, which is used to retrieve the input name, id, and values. Otherwise all attributes may be passed explicitly.

Examples

<.input field={@form[:email]} type="email" /> <.input name="my-input" errors={["oh no!"]} />

Attributes

id (:any) - Defaults to nil.

name (:any)

label (:string) - Defaults to nil.

value (:any)

type (:string) - Defaults to "text".

field (Phoenix.HTML.FormField) - a form field struct retrieved from the form, for example: @form[:email].

errors (:list) - Defaults to [].

checked (:boolean) - the checked flag for checkbox inputs.

prompt (:string) - the prompt for select inputs. Defaults to nil.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2.

multiple (:boolean) - the multiple flag for select inputs. Defaults to false.

Global attributes are accepted.

Slots

inner_block

label(assigns)

Renders a label.

Attributes

for (:string) - Defaults to nil.

Slots

inner_block (required)

translate_error(arg)