PlausibleWeb.Router.Helpers (Plausible v0.0.1)

Module with named helpers generated from PlausibleWeb.Router.

Functions

auth_path(conn_or_endpoint, action)

auth_path(conn_or_endpoint, action, params)

auth_path(conn_or_endpoint, action, id, params)

auth_url(conn_or_endpoint, action)

auth_url(conn_or_endpoint, action, params)

auth_url(conn_or_endpoint, action, id, params)

billing_path(conn_or_endpoint, action)

billing_path(conn_or_endpoint, action, params)

billing_path(conn_or_endpoint, action, plan_id, params)

billing_url(conn_or_endpoint, action)

billing_url(conn_or_endpoint, action, params)

billing_url(conn_or_endpoint, action, plan_id, params)

external_path(conn_or_endpoint, action)

external_path(conn_or_endpoint, action, params)

external_sites_path(conn_or_endpoint, action)

external_sites_path(conn_or_endpoint, action, params)

external_sites_path(conn_or_endpoint, action, goal_id, params)

external_sites_url(conn_or_endpoint, action)

external_sites_url(conn_or_endpoint, action, params)

external_sites_url(conn_or_endpoint, action, goal_id, params)

external_stats_path(conn_or_endpoint, action)

external_stats_path(conn_or_endpoint, action, params)

external_stats_url(conn_or_endpoint, action)

external_stats_url(conn_or_endpoint, action, params)

external_url(conn_or_endpoint, action)

external_url(conn_or_endpoint, action, params)

internal_path(conn_or_endpoint, action)

internal_path(conn_or_endpoint, action, params)

internal_path(conn_or_endpoint, action, domain, params)

internal_url(conn_or_endpoint, action)

internal_url(conn_or_endpoint, action, params)

internal_url(conn_or_endpoint, action, domain, params)

invitation_path(conn_or_endpoint, action, invitation_id)

invitation_path(conn_or_endpoint, action, invitation_id, params)

invitation_url(conn_or_endpoint, action, invitation_id)

invitation_url(conn_or_endpoint, action, invitation_id, params)

kaffy_api_resource_path(conn_or_endpoint, action, context, resource)

kaffy_api_resource_path(conn_or_endpoint, action, context, resource, params)

kaffy_api_resource_url(conn_or_endpoint, action, context, resource)

kaffy_api_resource_url(conn_or_endpoint, action, context, resource, params)

kaffy_dashboard_path(conn_or_endpoint, action)

kaffy_dashboard_path(conn_or_endpoint, action, params)

kaffy_dashboard_url(conn_or_endpoint, action)

kaffy_dashboard_url(conn_or_endpoint, action, params)

kaffy_home_path(conn_or_endpoint, action)

kaffy_home_path(conn_or_endpoint, action, params)

kaffy_home_url(conn_or_endpoint, action)

kaffy_home_url(conn_or_endpoint, action, params)

kaffy_page_path(conn_or_endpoint, action, slug)

kaffy_page_path(conn_or_endpoint, action, slug, params)

kaffy_page_url(conn_or_endpoint, action, slug)

kaffy_page_url(conn_or_endpoint, action, slug, params)

kaffy_resource_path(conn_or_endpoint, action, context, resource)

kaffy_resource_path(conn_or_endpoint, action, context, resource, params)

kaffy_resource_path(conn_or_endpoint, action, context, resource, id, params)

kaffy_resource_path(conn_or_endpoint, action, context, resource, id, action_key, params)

kaffy_resource_url(conn_or_endpoint, action, context, resource)

kaffy_resource_url(conn_or_endpoint, action, context, resource, params)

kaffy_resource_url(conn_or_endpoint, action, context, resource, id, params)

kaffy_resource_url(conn_or_endpoint, action, context, resource, id, action_key, params)

kaffy_task_path(conn_or_endpoint, action)

kaffy_task_path(conn_or_endpoint, action, params)

kaffy_task_url(conn_or_endpoint, action)

kaffy_task_url(conn_or_endpoint, action, params)

membership_path(conn_or_endpoint, action, website)

membership_path(conn_or_endpoint, action, website, params)

membership_path(conn_or_endpoint, action, website, id, params)

membership_path(conn_or_endpoint, action, website, id, new_role, params)

membership_url(conn_or_endpoint, action, website)

membership_url(conn_or_endpoint, action, website, params)

membership_url(conn_or_endpoint, action, website, id, params)

membership_url(conn_or_endpoint, action, website, id, new_role, params)

paddle_path(conn_or_endpoint, action)

paddle_path(conn_or_endpoint, action, params)

paddle_url(conn_or_endpoint, action)

paddle_url(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

site_path(conn_or_endpoint, action)

site_path(conn_or_endpoint, action, params)

site_path(conn_or_endpoint, action, website, params)

site_path(conn_or_endpoint, action, website, id, params)

site_url(conn_or_endpoint, action)

site_url(conn_or_endpoint, action, params)

site_url(conn_or_endpoint, action, website, params)

site_url(conn_or_endpoint, action, website, id, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

stats_path(conn_or_endpoint, action, slug)

stats_path(conn_or_endpoint, action, slug, params)

stats_path(conn_or_endpoint, action, domain, filter_name, params)

stats_url(conn_or_endpoint, action, slug)

stats_url(conn_or_endpoint, action, slug, params)

stats_url(conn_or_endpoint, action, domain, filter_name, params)

unsubscribe_path(conn_or_endpoint, action, website)

unsubscribe_path(conn_or_endpoint, action, website, params)

unsubscribe_url(conn_or_endpoint, action, website)

unsubscribe_url(conn_or_endpoint, action, website, params)

url(data)

Generates the connection/endpoint base URL without any path information.