PlausibleWeb.SiteView (Plausible v0.0.1)

Functions

__resource__()

The resource name, as an atom, for this view

change_domain.html(assigns)

csv_import.html(assigns)

edit_shared_link.html(assigns)

new.html(assigns)

new_shared_link.html(assigns)

plausible_url()

render(template, assigns \\ %{})

Renders the given template locally.

render_snippet(site)

settings_danger_zone.html(assigns)

settings_email_reports.html(assigns)

settings_funnels.html(assigns)

settings_general.html(assigns)

settings_goals.html(assigns)

settings_imports_exports.html(assigns)

settings_integrations.html(assigns)

settings_people.html(assigns)

settings_props.html(assigns)

settings_search_console.html(assigns)

settings_shields.html(assigns)

settings_visibility.html(assigns)

shared_link_dest(site, link)

template_not_found(template, assigns)

@spec template_not_found(binary(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

with_indefinite_article(word)