Plausible.Billing.Feature behaviour (Plausible v0.0.1)

This module provides an interface for managing features, e.g. Revenue Goals, Funnels and Custom Properties.

Feature modules have functions for toggling the feature on/off and checking whether the feature is available for a site/user.

When defining new features, the following options are expected by the __using__ macro:

:name - an atom representing the feature name in the plan JSON file (see also Plausible.Billing.Plan).

:display_name - human-readable display name of the feature

:toggle_field - the field in the %Plausible.Site{} schema that toggles the feature. If nil or not set, toggle/2 silently returns :ok

Functions defined by __using__ can be overridden if needed.

Types

t()

@type t() ::
  Plausible.Billing.Feature.ConsolidatedView
  | Plausible.Billing.Feature.SSO
  | Plausible.Billing.Feature.StatsAPI
  | Plausible.Billing.Feature.SitesAPI
  | Plausible.Billing.Feature.SiteSegments
  | Plausible.Billing.Feature.RevenueGoals
  | Plausible.Billing.Feature.Goals
  | Plausible.Billing.Feature.Funnels
  | Plausible.Billing.Feature.SharedLinks
  | Plausible.Billing.Feature.Props

Callbacks

check_availability(arg1)

@callback check_availability(Plausible.Teams.Team.t() | nil) ::
  :ok | {:error, :upgrade_required} | {:error, :not_implemented}

Checks whether the team or the team plan includes the given feature.

display_name()

@callback display_name() :: String.t()

Returns the human-readable display name of the feature.

enabled?(t)

@callback enabled?(Plausible.Site.t()) :: boolean()

Checks whether a feature is enabled or not. Returns false when the feature is disabled or the user does not have access to it.

name()

@callback name() :: atom()

Returns the atom representing the feature name in the plan JSON file.

opted_out?(t)

@callback opted_out?(Plausible.Site.t()) :: boolean()

Returns whether the site explicitly opted out of the feature. This function is different from enabled/1, because enabled/1 returns false when the site owner does not have access to the feature.

toggle(t, t, t)

@callback toggle(Plausible.Site.t(), Plausible.Auth.User.t(), Keyword.t()) ::
  :ok | {:error, :upgrade_required}

Toggles the feature on and off for a site. Returns {:error, :upgrade_required} when toggling a feature the site owner does not have access to.

toggle_field()

@callback toggle_field() :: atom()

Returns the %Plausible.Site{} field that toggles the feature on and off.

Functions

list()

Lists all available feature modules.

list_short_names()
(macro)

Lists all the feature short names, e.g. RevenueGoals