Plausible.Billing.Plans (Plausible v0.0.1)

Functions

all()

available_plans_for(subscription, opts \\ [])

business_plans_for(subscription)

business_tier?(subscription)

find(product_id)

get_price_for(enterprise_plan, customer_ip)

get_regular_plan(subscription, opts \\ [])

get_subscription_plan(subscription)

@spec get_subscription_plan(nil | Plausible.Billing.Subscription.t()) ::
  nil
  | :free_10k
  | Plausible.Billing.Plan.t()
  | Plausible.Billing.EnterprisePlan.t()

growth_plans_for(subscription)

@spec growth_plans_for(Plausible.Billing.Subscription.t()) :: [
  Plausible.Billing.Plan.t()
]

Returns a list of growth plans available for the subscription to choose.

As new versions of plans are introduced, subscriptions which were on old plans can still choose from old plans.

subscription_interval(subscription)

suggest_volume(team, usage_during_cycle)

@spec suggest_volume(Plausible.Teams.Team.t(), non_neg_integer()) ::
  String.t() | :enterprise

Returns the most appropriate monthly pageview volume for a given usage cycle. The cycle is either last 30 days (for trials) or last billing cycle for teams with an existing subscription.

The generation and tier from which we're searching for a suitable volume doesn't matter - the monthly pageview volumes for all plans starting from v3 are going from 10k to 10M. This function uses v4 Growth but it might as well be e.g. v5 Business.

If the usage during the cycle exceeds the enterprise-level threshold, or if the team already has an enterprise plan, it returns :enterprise. Otherwise, a string representing the volume, e.g. "100k" or "5M".

with_prices(plans, customer_ip \\ "127.0.0.1")

This function takes a list of plans as an argument, gathers all product IDs in a single list, and makes an API call to Paddle. After a successful response, fills in the monthly_cost and yearly_cost fields for each given plan and returns the new list of plans with completed information.

yearly_product_ids()

@spec yearly_product_ids() :: [String.t()]

List yearly plans product IDs.