Plausible.Billing.Quota (Plausible v0.0.1)

This module provides functions to work with plans usage and limits.

monthly_pageview_usage()

@type monthly_pageview_usage() :: %{required(period()) => usage_cycle()}

period()

@type period() :: :last_30_days | :current_cycle | :last_cycle | :penultimate_cycle

usage_cycle()

@type usage_cycle() :: %{
  date_range: Date.Range.t(),
  pageviews: non_neg_integer(),
  custom_events: non_neg_integer(),
  total: non_neg_integer()
}

allowed_features_for(user)

Returns a list of features the user can use. Trial users have the ability to use all features during their trial.

below_limit?(usage, limit)

@spec below_limit?(non_neg_integer(), non_neg_integer() | :unlimited) :: boolean()

Returns whether the usage is below the limit or not. Returns false if usage is equal to the limit.

ensure_can_subscribe_to_plan(user, plan, usage \\ nil)

features_usage(user)

@spec features_usage(Plausible.Auth.User.t() | Plausible.Site.t()) :: [atom()]

Given a user, this function returns the features used across all the sites this user owns + StatsAPI if the user has a configured Stats API key.

Given a site, returns the features used by the site.

monthly_pageview_limit(subscription)

@spec monthly_pageview_limit(Plausible.Billing.Subscription.t()) ::
  non_neg_integer() | :unlimited

Returns the limit of pageviews for a subscription.

monthly_pageview_usage(user)

@spec monthly_pageview_usage(Plausible.Auth.User.t()) :: monthly_pageview_usage()

site_limit(user)

@spec site_limit(Plausible.Auth.User.t()) :: non_neg_integer() | :unlimited

Returns the limit of sites a user can have.

For enterprise customers, returns :unlimited. The site limit is checked in a background job so as to avoid service disruption.

site_usage(user)

@spec site_usage(Plausible.Auth.User.t()) :: non_neg_integer()

Returns the number of sites the given user owns.

team_member_limit(user)

@spec team_member_limit(Plausible.Auth.User.t()) :: non_neg_integer()

Returns the limit of team members a user can have in their sites.

team_member_usage(user)

@spec team_member_usage(Plausible.Auth.User.t()) :: integer()

Returns the total count of team members and pending invitations associated with the user's sites.

usage(user, opts \\ [])

usage_cycle(user, cycle, today \\ Timex.today())

@spec usage_cycle(Plausible.Auth.User.t(), period(), Date.t()) :: usage_cycle()

within_limit?(usage, limit)

@spec within_limit?(non_neg_integer(), non_neg_integer() | :unlimited) :: boolean()

Returns whether the usage is within the limit or not. Returns true if usage is equal to the limit.