Plausible.Billing.Quota (Plausible v0.0.1)

This module provides functions to work with plans usage and limits.

monthly_pageview_limit(subscription)

@spec monthly_pageview_limit(Plausible.Billing.Subscription.t()) ::
  non_neg_integer() | :unlimited

Returns the limit of pageviews for a subscription.

monthly_pageview_usage(user)

@spec monthly_pageview_usage(Plausible.Auth.User.t()) :: non_neg_integer()

Returns the amount of pageviews sent by the sites the user owns in last 30 days.

site_limit(user)

@spec site_limit(Plausible.Auth.User.t()) :: non_neg_integer() | :unlimited

Returns the limit of sites a user can have.

For enterprise customers, returns :unlimited. The site limit is checked in a background job so as to avoid service disruption.

site_usage(user)

@spec site_usage(Plausible.Auth.User.t()) :: non_neg_integer()

Returns the number of sites the given user owns.

within_limit?(usage, limit)

@spec within_limit?(non_neg_integer(), non_neg_integer() | :unlimited) :: boolean()

Returns whether the limit has been exceeded or not.