Plausible.Billing.Quota (Plausible v0.0.1)

This module provides functions to work with plans usage and limits.

Types

cycle()

@type cycle() :: :current_cycle | :last_cycle | :penultimate_cycle

cycles_usage()

@type cycles_usage() :: %{required(cycle()) => usage_cycle()}

usage_cycle()

@type usage_cycle() :: %{
  date_range: Date.Range.t(),
  pageviews: non_neg_integer(),
  custom_events: non_neg_integer(),
  total: non_neg_integer()
}

Functions

below_limit?(usage, limit)

@spec below_limit?(non_neg_integer(), non_neg_integer() | :unlimited) :: boolean()

Returns whether the usage is below the limit or not. Returns false if usage is equal to the limit.

eligible_for_upgrade?(usage)

ensure_feature_access(usage, plan)

ensure_within_plan_limits(usage, plan_mod, opts \\ [])

@spec ensure_within_plan_limits(map(), struct() | atom() | nil, Keyword.t()) ::
  :ok | {:error, Plausible.Billing.Quota.Limits.over_limits_error()}

Ensures that the given usage map is within the limits of the given plan.

An opts argument can be passed with ignore_pageview_limit: true which bypasses the pageview limit check and returns :ok as long as the other limits are not exceeded.

exceeded_cycles(cycles_usage, allowed_volume)

@spec exceeded_cycles(cycles_usage(), non_neg_integer()) :: list()

exceeds_last_two_usage_cycles?(cycles_usage, allowed_volume)

@spec exceeds_last_two_usage_cycles?(cycles_usage(), non_neg_integer()) :: boolean()

suggest_tier(usage, highest_starter, highest_growth, highest_business, owned_tier)

Suggests a suitable tier (Starter, Growth or Business) for the given usage map.

If even the highest Business plan does not accommodate the usage, then :custom is returned. This means that this kind of usage should get on a custom plan.

To avoid confusion, we do not recommend a lower tier for customers that are already on a higher tier (even if their usage is low enough).

nil is returned if the usage is not eligible for upgrade.

within_limit?(usage, limit)

@spec within_limit?(non_neg_integer(), non_neg_integer() | :unlimited) :: boolean()

Returns whether the usage is within the limit or not. Returns true if usage is equal to the limit.