Plausible.Cache.Adapter (Plausible v0.0.1)

Interface for the underlying cache implementation. Currently: ConCache

Using the Adapter module directly, the user must ensure that the relevant processes are available to use, which is normally done via the child specification.

Functions

cache2list(full_cache_name)

child_spec(name, child_id, opts \\ [])

@spec child_spec(atom(), atom(), Keyword.t()) :: Supervisor.child_spec()

child_specs(name, child_id, opts \\ [])

@spec child_specs(atom(), atom(), Keyword.t()) :: [Supervisor.child_spec()]

delete(cache_name, key)

@spec delete(atom(), any()) :: :ok

fetch(cache_name, key, fallback_fn)

@spec fetch(atom(), any(), (-> any())) :: any()

get(cache_name, key)

@spec get(atom(), any()) :: any()

get(cache_name, key, fallback_fn)

@spec get(atom(), any(), (-> any())) :: any()

get_names(cache_name)

@spec get_names(atom()) :: [atom()]

keys(cache_name)

@spec keys(atom()) :: Enumerable.t()

put(cache_name, key, value, opts \\ [])

put_many(cache_name, items)

@spec put_many(atom(), [any()]) :: :ok

size(cache_name)

@spec size(atom()) :: non_neg_integer() | nil