Plausible.Cldr.Number.Symbol (Plausible v0.0.1)

Functions

all_decimal_symbols()

Returns a list of all decimal symbols defined by the locales configured in this backend as a list.

all_decimal_symbols_class()

Returns a list of all decimal symbols defined by the locales configured in this backend as a string.

This string can be used as a character class when builing a regular expression.

all_grouping_symbols()

Returns a list of all grouping symbols defined by the locales configured in this backend as a list.

all_grouping_symbols_class()

Returns a list of all grouping symbols defined by the locales configured in this backend as a string.

This string can be used as a character class when builing a regular expression.

number_symbols_for(locale \\ Plausible.Cldr.get_locale())

@spec number_symbols_for(Cldr.LanguageTag.t() | Cldr.Locale.locale_name()) ::
  {:ok, map()} | {:error, {module(), String.t()}}

Returns a map of Cldr.Number.Symbol.t structs of the number symbols for each of the number systems of a locale.

Options

locale is any valid locale name returned by Plausible.Cldr.known_locale_names/0 or a Cldr.LanguageTag struct returned by Plausible.Cldr.Locale.new!/1. The default is Plausible.Cldr.get_locale/0.

Example:

iex> Plausible.Cldr.Number.Symbol.number_symbols_for(:th)
{
  :ok,
  %{
    latn: %Cldr.Number.Symbol{
      decimal: %{standard: "."},
      exponential: "E",
      group: %{standard: ","},
      infinity: "∞",
      list: ";",
      minus_sign: "-",
      nan: "NaN",
      per_mille: "‰",
      percent_sign: "%",
      plus_sign: "+",
      superscripting_exponent: "×",
      time_separator: ":"
    },
    thai: %Cldr.Number.Symbol{
      decimal: %{standard: "."},
      exponential: "E",
      group: %{standard: ","},
      infinity: "∞",
      list: ";",
      minus_sign: "-",
      nan: "NaN",
      per_mille: "‰",
      percent_sign: "%",
      plus_sign: "+",
      superscripting_exponent: "×",
      time_separator: ":"
    }
  }
}

number_symbols_for(locale, number_system)