Plausible.Exports (Plausible v0.0.1)

Contains functions to export data for events and sessions as Zip archives.

Types

export()

@type export() :: %{
  path: Path.t(),
  name: String.t(),
  expires_at: DateTime.t() | nil,
  size: pos_integer()
}

Functions

archive_filename(domain, created_on)

Renders export archive filename.

Examples:

iex> archive_filename("plausible.io", _created_on = ~D[2024-12-31])
"plausible_io_20241231.zip"

content_disposition(filename)

Safely renders content disposition for an arbitrary export filename.

Examples:

iex> content_disposition("plausible_io_20241231.zip")
"attachment; filename=\"plausible_io_20241231.zip\""

iex> content_disposition("📊.zip")
"attachment; filename=\"plausible-export.zip\"; filename*=utf-8''%F0%9F%93%8A.zip"

date_range(site_id, timezone)

@spec date_range(pos_integer(), String.t()) :: Date.Range.t() | nil

Returns the date range for the site's events data in site's timezone or nil if there is no data

delete_local_export(site_id)

@spec delete_local_export(pos_integer()) :: :ok

Deletes local export for a site

delete_s3_export!(site_id)

@spec delete_s3_export!(pos_integer()) :: :ok

Deletes S3 export for a site. Raises if object storage is unavailable.

export_queries(site_id, opts \\ [])

@spec export_queries(pos_integer(),
  extname: String.t(),
  date_range: Date.Range.t(),
  timezone: String.t()
) :: %{required(String.t()) => Ecto.Query.t()}

Builds Ecto queries to export data from events_v2 and sessions_v2 tables into the format of imported_* tables for a website.

get_last_export_job(site_id)

@spec get_last_export_job(pos_integer()) :: Oban.Job.t() | nil

Gets last CSV export job for a site

get_local_export(site_id, domain, timezone)

@spec get_local_export(pos_integer(), String.t(), String.t()) :: export() | nil

Gets local export for a site

get_s3_export!(site_id, retries \\ 0)

@spec get_s3_export!(pos_integer(), non_neg_integer()) :: export() | nil

Gets S3 export for a site. Raises if object storage is unavailable.

oban_listen()

Subscribes to CSV export job notifications

schedule_local_export(site_id, email_to)

@spec schedule_local_export(pos_integer(), String.t()) ::
  {:ok, Oban.Job.t()} | {:error, :no_data}

Schedules CSV export job to local storage

schedule_s3_export(site_id, email_to)

@spec schedule_s3_export(pos_integer(), String.t()) ::
  {:ok, Oban.Job.t()} | {:error, :no_data}

Schedules CSV export job to S3 storage

stream_archive(conn, named_queries, opts \\ [])

@spec stream_archive(DBConnection.t(), %{required(String.t()) => Ecto.Query.t()}, [
  Ch.query_option()
]) ::
  Enumerable.t()

Creates a streamable Zip archive from the provided (named) Ecto queries.

Example usage:

{:ok, pool} = Ch.start_link(pool_size: 1)

DBConnection.run(pool, fn conn ->
  conn
  |> stream_archive(export_queries(_site_id = 1), format: "CSVWithNames")
  |> Stream.into(File.stream!("export.zip"))
  |> Stream.run()
end)