Plausible.Funnels (Plausible v0.0.1)

This module implements contextual Funnel interface, allowing listing, creating and deleting funnel definitions.

For brief explanation of what a Funnel is, please see Plausible.Funnel schema. See Plausible.Stats.Funnel for the evaluation logic.

Functions

create(site, name, steps)

@spec create(Plausible.Site.t(), String.t(), [map()]) ::
  {:ok, Plausible.Funnel.t()}
  | {:error, Ecto.Changeset.t() | :invalid_funnel_size | :upgrade_required}

create_changeset(site, name, steps)

@spec create_changeset(Plausible.Site.t(), String.t(), [map()]) :: Ecto.Changeset.t()

delete(site_id, funnel_id)

@spec delete(Plausible.Site.t() | pos_integer(), pos_integer()) :: :ok

edit_changeset(funnel, name, steps)

@spec edit_changeset(Plausible.Funnel.t(), String.t(), [map()]) :: Ecto.Changeset.t()

ephemeral_definition(site, name, steps)

@spec ephemeral_definition(Plausible.Site.t(), String.t(), [map()]) ::
  Plausible.Funnel.t()

get(site_id, by)

@spec get(Plausible.Site.t() | pos_integer(), pos_integer() | String.t()) ::
  Plausible.Funnel.t() | nil

list(site)

@spec list(Plausible.Site.t()) :: [
  %{name: String.t(), id: pos_integer(), steps_count: pos_integer()}
]

update(funnel, name, steps)

@spec update(Plausible.Funnel.t(), String.t(), [map()]) ::
  {:ok, Plausible.Funnel.t()}
  | {:error, Ecto.Changeset.t() | :invalid_funnel_size | :upgrade_required}

with_goals_query(site)

@spec with_goals_query(Plausible.Site.t()) :: Ecto.Query.t()