Plausible.Goals (Plausible v0.0.1)

Functions

batch_create_event_goals(names, site, opts \\ [])

count(site)

@spec count(Plausible.Site.t()) :: non_neg_integer()

create(site, params, opts \\ [])

@spec create(Plausible.Site.t(), map(), Keyword.t()) ::
  {:ok, Plausible.Goal.t()}
  | {:error, Ecto.Changeset.t()}
  | {:error, :upgrade_required}
  | {:error, :revenue_goals_unavailable}

Creates a Goal for a site.

If the created goal is a revenue goal, it sets site.updated_at to be refreshed by the sites cache, as revenue goals are used during ingestion.

Returns {:ok, goal} or {:error, changeset} when creation fails due to invalid fields. It can also return:

{:error, :upgrade_required} - Adding a revenue goal is not allowed for team's subscription.

{:error, :revenue_goals_unavailable} - When the site is a consolidated view and the goal created is a revenue goal. Revenue goal creation is not allowed for consolidated views due to the inability to force a single currency on a goal across all consolidated sites.

create_404(site)

@spec create_404(Plausible.Site.t()) :: :ok

create_file_downloads(site)

@spec create_file_downloads(Plausible.Site.t()) :: :ok

create_form_submissions(site)

@spec create_form_submissions(Plausible.Site.t()) :: :ok

create_outbound_links(site)

@spec create_outbound_links(Plausible.Site.t()) :: :ok

delete(id, site_id)

If a goal belongs to funnel(s), we need to inspect their number of steps.

If it exceeds the minimum allowed (defined via Plausible.Funnel.min_steps/0), the funnel will be reduced (i.e. a step associated with the goal to be deleted is removed), so that the minimum number of steps is preserved. This is done implicitly, by postgres, as per on_delete: :delete_all.

Otherwise, for associated funnel(s) consisting of minimum number steps only, funnel record(s) are removed completely along with the targeted goal.

delete_404(site)

@spec delete_404(Plausible.Site.t()) :: :ok

delete_file_downloads(site)

@spec delete_file_downloads(Plausible.Site.t()) :: :ok

delete_form_submissions(site)

@spec delete_form_submissions(Plausible.Site.t()) :: :ok

delete_outbound_links(site)

@spec delete_outbound_links(Plausible.Site.t()) :: :ok

find_or_create(site, params, opts \\ [])

for_site(site, opts \\ [])

for_site_query(site \\ nil, opts \\ [])

get(site, id)

@spec get(Plausible.Site.t(), pos_integer()) :: nil | Plausible.Goal.t()

list_revenue_goals(site)

max_goals_per_site(opts \\ [])

@spec max_goals_per_site(Keyword.t()) :: pos_integer()

update(goal, params)

@spec update(Plausible.Goal.t(), map()) ::
  {:ok, Plausible.Goal.t()}
  | {:error, Ecto.Changeset.t()}
  | {:error, :upgrade_required}