Plausible.Ingestion.Request (Plausible v0.0.1)

The %Plausible.Ingestion.Request{} struct stores all needed fields to create an event downstream. Pre-eliminary validation is made to detect user errors early.

Types

t()

@type t() :: %Plausible.Ingestion.Request{
  domains: term(),
  engagement_time: term(),
  event_name: term(),
  hash_mode: term(),
  hostname: term(),
  interactive?: term(),
  ip_classification: term(),
  pathname: term(),
  props: term(),
  query_params: term(),
  referrer: term(),
  remote_ip: term(),
  revenue_source: term(),
  scroll_depth: term(),
  timestamp: term(),
  tracker_script_version: term(),
  uri: term(),
  user_agent: term()
}

Functions

blank_engagement_error_message()

build(conn, now \\ NaiveDateTime.utc_now())

@spec build(Plug.Conn.t(), NaiveDateTime.t()) ::
  {:ok, t(), Plug.Conn.t()} | {:error, Ecto.Changeset.t()}

Builds and initially validates %Plausible.Ingestion.Request{} struct from %Plug.Conn{}.

sanitize_hostname(hostname)

Removes the "www" part of a hostname.

too_large_engagement_time()