Plausible.Sites (Plausible v0.0.1)

create(user, params)

create_shared_link(site, name, password \\ nil)

get_by_domain(domain)

get_for_user(user_id, domain, roles \\ [:owner, :admin, :viewer])

get_for_user!(user_id, domain, roles \\ [:owner, :admin, :viewer])

has_admin_access?(user_id, site)

has_goals?(site)

has_stats?(site)

is_member?(user_id, site)

locked?(site)

owned_site_ids(user)

owned_sites_count(user)

owned_sites_domains(user)

owner_for(site)

role(user_id, site)

set_allowed_event_props!(site, props)

shared_link_url(site, link)

stats_start_date(site)

@spec stats_start_date(Plausible.Site.t()) :: Date.t() | nil

Returns the date of the first event of the given site, or nil if the site does not have stats yet.

If this is the first time the function is called for the site, it queries Clickhouse and saves the date in the sites table.