Plausible.Sites (Plausible v0.0.1)

Sites context functions.

Functions

clear_stats_start_date!(site)

@spec clear_stats_start_date!(Plausible.Site.t()) :: Plausible.Site.t()

create(user, params)

create_shared_link(site, name, password \\ nil)

for_user_query(user)

@spec for_user_query(Plausible.Auth.User.t()) :: Ecto.Query.t()

get_by_domain(domain)

get_by_domain!(domain)

get_for_user(user, domain, roles \\ [:owner, :admin, :viewer])

get_for_user!(user, domain, roles \\ [:owner, :admin, :viewer])

has_goals?(site)

has_stats?(site)

list(user, pagination_params, opts \\ [])

See Plausible.Teams.Sites.list/3.

list_people(site)

list_with_invitations(user, pagination_params, opts \\ [])

See Plausible.Teams.Sites.list_with_invitations/3.

locked?(site)

native_stats_start_date(site)

@spec native_stats_start_date(Plausible.Site.t()) :: Date.t() | nil

set_option(user, site, option, value)

@spec set_option(Plausible.Auth.User.t(), Plausible.Site.t(), atom(), any()) ::
  Plausible.Site.UserPreference.t()

set_scroll_depth_visible_at(site)

shared_link_url(site, link)

stats_start_date(site)

@spec stats_start_date(Plausible.Site.t()) :: Date.t() | nil

Returns the date of the first event of the given site, or nil if the site does not have stats yet.

If this is the first time the function is called for the site, it queries imported stats and Clickhouse, choosing the earliest start date and saves it in the sites table.

toggle_pin(user, site)

@spec toggle_pin(Plausible.Auth.User.t(), Plausible.Site.t()) ::
  {:ok, Plausible.Site.UserPreference.t()} | {:error, :too_many_pins}

update_installation_meta!(site, meta)