Plausible.Sites (Plausible v0.0.1)

Sites context functions.

Functions

clear_stats_start_date!(site)

@spec clear_stats_start_date!(Plausible.Site.t()) :: Plausible.Site.t()

consolidated?(site)

@spec consolidated?(Plausible.Site.t()) :: boolean()

create(user, params, team \\ nil)

create_shared_link(site, name, opts \\ [])

display_name(site, opts \\ [])

for_user_query(user, team \\ nil)

@spec for_user_query(Plausible.Auth.User.t(), Plausible.Teams.Team.t() | nil) ::
  Ecto.Query.t()

get_by_domain(domain, opts \\ [])

get_by_domain!(domain, opts \\ [])

get_for_user(user, domain, opts \\ [])

get_for_user!(user, domain, opts \\ [])

has_goals?(site)

has_stats?(site)

list(user, pagination_params, opts \\ [])

See Plausible.Teams.Sites.list/3.

list_guests_query(site, opts \\ [])

@spec list_guests_query(Plausible.Site.t(), Keyword.t()) :: Ecto.Query.t()

list_people(site)

list_with_invitations(user, pagination_params, opts \\ [])

See Plausible.Teams.Sites.list_with_invitations/3.

native_stats_start_date(site)

@spec native_stats_start_date(Plausible.Site.t()) :: Date.t() | nil

regular?(site)

@spec regular?(Plausible.Site.t()) :: boolean()

set_option(user, site, option, value)

@spec set_option(Plausible.Auth.User.t(), Plausible.Site.t(), atom(), any()) ::
  Plausible.Site.UserPreference.t()

shared_link_special_names()

Special shared link names are used to distinguish between those created by the Plugins API, and those created in any other way (i.e. via Sites API or in the Dashboard Site Settings UI).

The intent is to give our WP plugin the ability to display an embedded dashboard even when the user's subscription does not support the shared links feature.

A shared link with a special name can only be created via the plugins API, and it will not show up under the list of shared links in Site Settings > Visibility.

Once created with the special name, the link will be accessible even when the team does not have access to SharedLinks feature.

shared_link_url(site, link)

site_id_query_filter(site)

stats_start_date(site)

@spec stats_start_date(Plausible.Site.t()) :: Date.t() | nil

Returns the date of the first event of the given site, or nil if the site does not have stats yet.

If this is the first time the function is called for the site, it queries imported stats and Clickhouse, choosing the earliest start date and saves it in the sites table.

toggle_pin(user, site)

@spec toggle_pin(Plausible.Auth.User.t(), Plausible.Site.t()) ::
  {:ok, Plausible.Site.UserPreference.t()} | {:error, :too_many_pins}

update_legacy_time_on_page_cutoff!(site, cutoff)