Plausible.Stats.Filters (Plausible v0.0.1)

A module for parsing filters used in stat queries.

Functions

all_leaf_filters(filters)

dimensions_used_in_filters(filters, opts \\ [])

event_props()

event_table_visit_props()

filtering_on_dimension?(query, dimension, opts \\ [])

get_toplevel_filter(query, prefix)

Gets the first top level filter with matching dimension (or nil).

Only use in cases where it's known that filters are only set on the top level as it does not handle AND/OR/NOT!

parse(filters)

Parses different filter formats.

Depending on the format and type of the filters argument, returns:

a parsed filter list, when filters is a filter expression string

the same list, when filters is a map

Returns an empty list when argument type is unexpected (e.g. nil).

Examples:

iex> Filters.parse("visit:browser!=Chrome")
[[:is_not, "visit:browser", ["Chrome"]]]

iex> Filters.parse(nil)
[]

rename_dimensions_used_in_filter(filters, renames)

transform_filters(filters, transformer)

Updates filters via transformer.

Transformer will receive each node (filter, and/or/not subtree) of query and must return a list of nodes to replace it with or nil to ignore and look deeper.

Examples

iex> Filters.transformfilters([[:is, "visit:os", ["Linux"]], [:and, [[:is, "segment", [1]], [:is, "segment", [2]]]]], fn ...> [, "segment", ] -> [[:is, "segment", ["changed"]]] ...> -> nil ...> end) [[:is, "visit:os", ["Linux"]], [:and, [[:is, "segment", ["changed"]], [:is, "segment", ["changed"]]]]]

traverse(filters, state \\ nil, state_transformer \\ fn state, _ -> state end)

Traverses a filter tree while accumulating state.

visit_props()

without_prefix(dimension)