PlausibleWeb.Email (Plausible v0.0.1)

activation_email(user, code)

base_email()

base_email(map)

cancellation_email(user)

check_stats_email(user)

create_site_email(user)

dashboard_locked(user, usage, last_cycle, suggested_plan)

enterprise_over_limit_internal_email(user, usage, last_cycle, site_usage, site_allowance)

error_report(reported_by, trace_id, feedback)

existing_user_invitation(invitation)

import_failure(user, site)

import_success(user, site)

invitation_accepted(invitation)

invitation_rejected(invitation)

mailer_email_from()

new_user_invitation(invitation)

over_limit_email(user, usage, last_cycle, suggested_plan)

ownership_transfer_accepted(invitation)

ownership_transfer_rejected(invitation)

ownership_transfer_request(invitation, new_owner_account)

password_reset_email(email, reset_link)

render(email, template, assigns \\ [])

Render an Phoenix template and set the body on the email.

Pass an atom as the template name (:welcome_email) to render HTML and plain text emails. Use a string if you only want to render one type, e.g. "welcome_email.text" or "welcome_email.html". Scroll to the top for more examples.

site_member_removed(membership)

site_setup_help(user, site)

site_setup_success(user, site)

spike_notification(email, site, current_visitors, sources, dashboard_link)

trial_one_week_reminder(user)

trial_over_email(user)

trial_upgrade_email(user, day, arg)

weekly_report(email, site, assigns)

welcome_email(user)

yearly_expiration_notification(user)

yearly_renewal_notification(user)