PlausibleWeb.Email (Plausible v0.0.1)

Functions

activation_email(user, code)

approaching_accept_traffic_until(notification)

approaching_accept_traffic_until_tomorrow(notification)

base_email()

base_email(map)

cancellation_email(user)

check_stats_email(user)

create_site_email(user)

dashboard_locked(user, usage, suggested_plan)

drop_notification(email, site, current_visitors, dashboard_link, installation_link)

enterprise_over_limit_internal_email(user, pageview_usage, site_usage, site_allowance)

error_report(reported_by, trace_id, feedback)

existing_user_invitation(email, site, inviter)

export_failure(user, site)

export_success(user, site, expires_at)

import_failure(site_import, user)

import_success(site_import, user)

invitation_accepted(inviter_email, invitee_email, site)

invitation_rejected(invitation)

mailer_email_from()

new_user_invitation(email, invitation_id, site, inviter)

over_limit_email(user, usage, suggested_plan)

ownership_transfer_accepted(new_owner_email, inviter_email, site)

ownership_transfer_rejected(invitation)

ownership_transfer_request(email, invitation_id, site, inviter, new_owner_account)

password_reset_email(email, reset_link)

priority_email()

Unlike the default 'base' emails, priority emails cannot be unsubscribed from. This is achieved by sending them through a dedicated 'priority' message stream in Postmark.

priority_email(map)

render(email, template, assigns \\ [])

Render an Phoenix template and set the body on the email.

Pass an atom as the template name (:welcome_email) to render HTML and plain text emails. Use a string if you only want to render one type, e.g. "welcome_email.text" or "welcome_email.html". Scroll to the top for more examples.

site_member_removed(membership)

site_setup_help(user, site)

site_setup_success(user, site)

spike_notification(email, site, current_visitors, sources, dashboard_link)

stats_report(email, assigns)

trial_one_week_reminder(user)

trial_over_email(user)

trial_upgrade_email(user, day, usage)

two_factor_disabled_email(user)

two_factor_enabled_email(user)

welcome_email(user)

yearly_expiration_notification(user)

yearly_renewal_notification(user)