PlausibleWeb.EmailView (Plausible v0.0.1)

Functions

__resource__()

The resource name, as an atom, for this view

activation_email.html(assigns)

approaching_accept_traffic_until.html(assigns)

cancellation_email.html(assigns)

check_stats_email.html(assigns)

create_site_email.html(assigns)

csv_import.html(assigns)

dashboard_locked.html(assigns)

date_format(date)

drop_notification.html(assigns)

enterprise_over_limit_internal.html(assigns)

error_report_email.html(assigns)

existing_user_invitation.html(assigns)

existing_user_team_invitation.html(assigns)

export_failure.html(assigns)

export_success.html(assigns)

force_2fa_enabled.html(assigns)

google_analytics_import.html(assigns)

greet_recipient(_)

guest_invitation_accepted.html(assigns)

guest_invitation_rejected.html(assigns)

guest_to_team_member_promotion.html(assigns)

new_user_invitation.html(assigns)

new_user_team_invitation.html(assigns)

over_limit.html(assigns)

ownership_transfer_accepted.html(assigns)

ownership_transfer_rejected.html(assigns)

ownership_transfer_request.html(assigns)

password_reset_email.html(assigns)

plausible_url()

render(template, assigns \\ %{})

Renders the given template locally.

sentry_link(trace_id, dsn \\ Sentry.Config.dsn())

site_member_removed.html(assigns)

site_setup_help_email.html(assigns)

site_setup_success_email.html(assigns)

spike_notification.html(assigns)

sso_domain_verification_failure.html(assigns)

sso_domain_verification_success.html(assigns)

team_changed.html(assigns)

team_invitation_accepted.html(assigns)

team_invitation_rejected.html(assigns)

team_member_left.html(assigns)

team_member_removed.html(assigns)

template_not_found(template, assigns)

@spec template_not_found(binary(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

trial_one_week_reminder.html(assigns)

trial_over_email.html(assigns)

trial_upgrade_email.html(assigns)

two_factor_disabled_email.html(assigns)

two_factor_enabled_email.html(assigns)

welcome_email.html(assigns)

yearly_expiration_notification.html(assigns)

yearly_renewal_notification.html(assigns)