PlausibleWeb.LayoutView (Plausible v0.0.1)

Functions

__resource__()

The resource name, as an atom, for this view

account_settings_sidebar(conn)

app.html(assigns)

base_email.html(assigns)

base_error.html(assigns)

embedded.html(assigns)

feedback_link(user)

grace_period_end(user)

home_dest(conn)

is_current_tab(conn, tab)

logo_path(filename)

plausible_url()

priority_email.html(assigns)

render(template, assigns \\ %{})

Renders the given template locally.

render_layout(layout, assigns, list)

http://blog.plataformatec.com.br/2018/05/nested-layouts-with-phoenix/

settings.html(assigns)

site_settings.html(assigns)

site_settings_sidebar(conn)

team_switcher(assigns)

Attributes

conn (:map) (required)

teams (:list) (required)

my_team (:any) - Defaults to nil.

current_team (:any) - Defaults to nil.

more_teams? (:boolean) (required)

template_not_found(template, assigns)

@spec template_not_found(binary(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.

trial_notification(team)

websocket_url()