PlausibleWeb.Live.Components.Form (Plausible v0.0.1)

Generic components stolen from mix phx.new templates

Functions

error(assigns)

Generates a generic error message.

Slots

inner_block (required)

input(assigns)

Renders an input with label and error messages.

A Phoenix.HTML.FormField may be passed as argument, which is used to retrieve the input name, id, and values. Otherwise all attributes may be passed explicitly.

Examples

<.input field={@form[:email]} type="email" /> <.input name="my-input" errors={["oh no!"]} />

Attributes

id (:any) - Defaults to nil.

name (:any)

label (:string) - Defaults to nil.

help_text (:string) - Defaults to nil.

value (:any)

width (:string) - Defaults to "w-full".

type (:string) - Defaults to "text". Must be one of "checkbox", "color", "date", "datetime-local", "email", "file", "hidden", "month", "number", "password", "range", "radio", "search", "select", "tel", "text", "textarea", "time", "url", or "week".

field (Phoenix.HTML.FormField) - a form field struct retrieved from the form, for example: @form[:email].

errors (:list) - Defaults to [].

checked (:boolean) - the checked flag for checkbox inputs.

prompt (:string) - the prompt for select inputs. Defaults to nil.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2.

multiple (:boolean) - the multiple flag for select inputs. Defaults to false.

class (:any) - Defaults to "text-sm text-gray-900 dark:text-white dark:bg-gray-750 block pl-3.5 py-2.5 border-gray-300 dark:border-gray-800 transition-all duration-150 focus:outline-none focus:ring-3 focus:ring-indigo-500/20 dark:focus:ring-indigo-500/25 focus:border-indigo-500 rounded-md disabled:bg-gray-100 disabled:dark:bg-gray-800 disabled:border-gray-200 disabled:dark:border-gray-800 disabled:text-gray-900/40 disabled:dark:text-white/30 disabled:cursor-not-allowed".

mt? (:boolean) - Defaults to true.

max_one_error (:boolean) - Defaults to false.

Global attributes are accepted. Supports all globals plus: ["accept", "autocomplete", "capture", "cols", "disabled", "form", "list", "max", "maxlength", "min", "minlength", "multiple", "pattern", "placeholder", "readonly", "required", "rows", "size", "step", "x-model"].

Slots

help_content

inner_block

input_with_clipboard(assigns)

Attributes

id (:string) (required)

name (:string) (required)

label (:string) - Defaults to nil.

value (:string) - Defaults to "".

Global attributes are accepted.

label(assigns)

Renders a label.

Attributes

for (:string) - Defaults to nil.

class (:string) - Defaults to "".

Slots

inner_block (required)

mobile_nav_dropdown(assigns)

Attributes

conn (Plug.Conn) (required)

name (:string) (required)

options (:list) (required)

value (:any) - Defaults to nil.

href_base (:string) - Defaults to "/".

selected_fn (:any) (required)

password_input_with_strength(assigns)

Attributes

id (:any) - Defaults to nil.

label (:string) - Defaults to nil.

field (Phoenix.HTML.FormField) (required) - a form field struct retrieved from the form, for example: @form[:password].

strength (:any)

Global attributes are accepted. Supports all globals plus: ["autocomplete", "disabled", "form", "maxlength", "minlength", "readonly", "required", "size"].

password_length_hint(assigns)

Attributes

minimum (:integer) (required)

class (:any)

ok_class (:any)

error_class (:any)

field (Phoenix.HTML.FormField) (required) - a form field struct retrieved from the form, for example: @form[:password].

strength_meter(assigns)

Attributes

score (:integer) - Defaults to 0.

warning (:string) - Defaults to "".

suggestions (:list) - Defaults to [].

translate_error(arg)